requirements.txt
colorama
google-generativeai
pdfplumber
python-dotenv